spanset: assert read-only batches don't access store local and unreplicated RangeID local keys
Similar to what we did in #157153, this commit does the same
but for read-only batches.

Testing showed no failures.
